+++ .gitignore
@@ -1,1 +1,3 @@
 *.log
+NFT_SNiper.egg-info/
+build/
