+++ useful_links.txt
@@ -39,3 +39,6 @@ https://docs.opensea.io/reference/request-an-api-key
 
 Finding NFTs On IPFS:
 https://medium.com/coinmonks/how-to-find-your-nft-on-ipfs-e51bc5e7c8a1
+
+NFTGO Rarity Calculation Model (Dec 2021)
+https://nftgo.medium.com/the-ultimate-guide-to-nftgos-new-rarity-model-3f2265dd0e23
